+++ upkie/cpp/observers/observe_servos.h
@@ -18,11 +18,11 @@ using upkie::cpp::actuation::moteus::ServoReply;
 /*! Observe servo measurements.
  *
  * \param[out] observation Dictionary to write observations to.
- * \param[in] servo_joint_map Map from servo ID to joint name.
+ * \param[in] servo_name_map Map from servo ID to joint name.
  * \param[in] servo_replies List of servo replies from the CAN bus.
  */
 void observe_servos(palimpsest::Dictionary& observation,
-                    const std::map<int, std::string>& servo_joint_map,
+                    const std::map<int, std::string>& servo_name_map,
                     const std::vector<ServoReply>& servo_replies);
 
 }  // namespace upkie::cpp::observers